move components in correct place and add custom Link
