+++ src/components/common/link/Link.tsx
@@ -1,8 +1,8 @@
-import { Link, LinkProps } from 'react-router-dom';
+import { Link as RouterLink, LinkProps } from 'react-router-dom';
 import { useLocale } from '@/hooks/common/use-locale/useLocale';
 import { DEFAULT_LOCALE } from '@/const/common/locales';
 
-export const LocalizedLink = ({ to, children, ...props }: LinkProps) => {
+export const Link = ({ to, children, ...props }: LinkProps) => {
     const { currentLanguage } = useLocale();
 
     const getLocalizedPath = (path: string) => {
@@ -12,12 +12,12 @@ export const LocalizedLink = ({ to, children, ...props }: LinkProps) => {
 
         if (path.startsWith(`/${currentLanguage}`)) return path;
 
-        return `/${currentLanguage}${path.startsWith('/') ? path : `/${path}`}`;
+        return `/${currentLanguage}${path.startsWith('/') ? path : `${path}`}`;
     };
 
     return (
-        <Link {...props} to={getLocalizedPath(to as string)}>
+        <RouterLink {...props} to={getLocalizedPath(to as string)}>
             {children}
-        </Link>
+        </RouterLink>
     );
 };
